+++ src/modules/router.module.js
@@ -5,12 +5,69 @@ pyscript.defmodule('router')
         self._params = {};
         self._promises = [];
 
+        self.proxy = {
+            setHash: function(hash) {
+                window.location.hash = hash;
+            },
+            setHref: function(href) {
+                window.location.href = href;
+            },
+            setPathname: function(path) {
+                window.location.pathname = path;
+            },
+            getHash: function() {
+                return window.location.hash;
+            },
+            getHref: function() {
+                return window.location.href;
+            },
+            getPathname: function() {
+                return window.location.pathname;
+            }
+        };
+
         window.addEventListener("hashchange", function() {
             self._onchange.call(self);
         });
     })
 
     .def({
+        mockSetup: function(self, throwErrorOnRefresh) {
+            pyscript.assert(jasmine, "mockSetup() can only be called in Jasmine testing.");
+
+            spyOn(self.proxy, 'setHash').and.callFake(function(value) {
+                self.proxy.setHref(self.proxy.getHref().split('#')[0] + '#' + value);
+                pyscript.defer(self._onchange);
+            });
+
+            spyOn(self.proxy, 'getHash').and.callFake(function() {
+                var hash = self.proxy.getHref().split('#')[1];
+                return hash ? ('#' + hash) : '';
+            });
+
+            spyOn(self.proxy, 'setHref').and.callFake(function(value) {
+                if (throwErrorOnRefresh)
+                    throw new Error('Page refresh detected. Redirection to: ' + value);
+                self.proxy._href = value;
+            });
+
+            spyOn(self.proxy, 'getHref').and.callFake(function() {
+                return self.proxy._href || window.location.href;
+            });
+
+            spyOn(self.proxy, 'setPathname').and.callFake(function(value) {
+                if (throwErrorOnRefresh)
+                    throw new Error('Page refresh detected. Redirection to: ' + value);
+                self.proxy.setHref(self.proxy.getHref().replace(self.proxy.getPathname(), value));
+            });
+
+            spyOn(self.proxy, 'getPathname').and.callFake(function() {
+                var href = self.proxy.getHref();
+                var pathname = href.replace('://', '').split('/', 1)[1] || "";
+                pathname = pathname.split('?')[0].split('#')[0];
+                return pathname;
+            });
+        },
         refresh: function() {
             pyscript.defer(function() {
                 var event = document.createEvent('Event');
@@ -28,7 +85,7 @@ pyscript.defmodule('router')
             return self;
         },
         _onchange: function (self) {
-            var paths = window.location.hash.slice(2).split('?')[0].split("/");
+            var paths = self.proxy.getHash().slice(2).split('?')[0].split("/");
 
             var queryParams = self.parseQuery();
             var route = "";
@@ -50,8 +107,8 @@ pyscript.defmodule('router')
             // Clear resolved promises
             self._promises = [];
         },
-        parseQuery: function() {
-            var hash = window.location.hash;
+        parseQuery: function(self) {
+            var hash = self.proxy.getHash();
             var query = [];
             if (hash.indexOf("?")) {
                 query = hash.slice(2).split("?");
@@ -78,7 +135,8 @@ pyscript.defmodule('router')
             pyscript.check(uri, String);
             var async = pyscript.async();
             self._promises.push(async);
-            window.location.hash = uri + this.asQueryString(self._params);
+
+            self.proxy.setHash(uri + this.asQueryString(self._params));
 
             if (force)
                 self.refresh();
@@ -90,10 +148,10 @@ pyscript.defmodule('router')
             self._params = params;
             var queryParams = self.parseQuery();
             pyscript.extend(queryParams, params);
-            window.location.href = window.location.href.split("?")[0] + self.asQueryString(self._params);
+            self.proxy.setHref(self.proxy.getHref().split("?")[0] + self.asQueryString(self._params));
         },
         redirect: function(self, pathname) {
-            window.location.pathname = pathname;
+            self.proxy.setPathname(pathname);
         }
     });