+++ tests/modules/cache.spec.js
@@ -73,12 +73,32 @@ describe('cache.module', function () {
             });
         });
 
-        it('should fetch values from server', function() {
-            pyscript.cache.flush();
+        it('should fetch values from server', function(done) {
             pyscript.requests.mockSetup();
-            pyscript.cache.fetch("keyZ");
+
+            pyscript.requests.mockServer.defRoute("GET", "keyZ", function() {
+                return {responseText: "Awesome", http: {success: true}};
+            });
+
+            pyscript.cache.fetch("keyZ",
+                function(value) { return value + ":Parsed"; })
+                .then(function(context) {
+                    expect(context.result).toBe("Awesome:Parsed");
+                    done();
+                });
 
             expect(pyscript.requests.get).toHaveBeenCalledWith('keyZ');
+        });
+
+
+        it('should upload file when calling syncFile()', function() {
+            pyscript.requests.mockSetup();
+
+            pyscript.cache.store("keyZ", {url: "awesomeUrl", file: {file: "fileData"}});
+            pyscript.cache.syncFile("keyZ");
+
+            expect(pyscript.requests.upload).toHaveBeenCalledWith(
+                'awesomeUrl', "fileData", {file: "fileData"});
         });
     });
 });